Handle layer selection update when changing biodiversity variant
