+++ src/components/data-global-sidebar/biodiversity-sidebar-card/biodiversity-sidebar-card.js
@@ -8,16 +8,18 @@ import Component from './biodiversity-sidebar-card-component';
 import { LAYERS_CATEGORIES, layersConfig } from 'constants/mol-layers-configs';
 import { batchToggleLayers, layerManagerToggle, flyToLayerExtent } from 'utils/layer-manager-utils';
 import mapStateToProps from './biodiversity-sidebar-card-selectors';
-import { biodiversityCategories } from 'constants/mol-layers-configs';
-import { useSelectLayersOnTabChange } from './biodiversity-sidebar-card-hooks';
-import { BIODIVERSITY_TABS_SLUGS } from 'constants/ui-params';
+
+import usePrevious from 'hooks/use-previous';
 import { ALL_TAXA_PRIORITY } from 'constants/layers-slugs';
-import { LAYERS_RESOLUTION, TERRESTRIAL, DEFAULT_RESOLUTION } from 'constants/biodiversity-layers-constants';
+import { LAYERS_RESOLUTION, LAYERS_TOGGLE_CONFIG, LAYER_VARIANTS, TERRESTRIAL, DEFAULT_RESOLUTION } from 'constants/biodiversity-layers-constants';
 
 const actions = {...metadataActions, ...urlActions};
 const BiodiversitySidebarCard = (props)  => {
   const { changeGlobe, changeUI, activeLayers, biodiversityLayerVariant, view } = props;
-  const { PRIORITY, RICHNESS, RARITY } = BIODIVERSITY_TABS_SLUGS;
+  const { PRIORITY, RICHNESS, RARITY } = LAYER_VARIANTS;
+  const previousBiodiversityLayerVariant = usePrevious(
+    biodiversityLayerVariant
+  );
   const [cardMetadata, setCardMetadata] = useState({
     [PRIORITY] : {},
     [RICHNESS] : {},
@@ -48,6 +50,29 @@ const BiodiversitySidebarCard = (props)  => {
     } 
   }, [biodiversityLayerVariant]);
 
+  useEffect(() => {
+    if (!previousBiodiversityLayerVariant) return;
+    const activeBiodiversityLayers = activeLayers
+    .filter((l) => l.category === LAYERS_CATEGORIES.BIODIVERSITY)
+    .map((l) => l.title);
+    const resolution = selectedResolution[TERRESTRIAL];
+    const deafaultResolutionLayers = LAYERS_TOGGLE_CONFIG[biodiversityLayerVariant][TERRESTRIAL][DEFAULT_RESOLUTION[TERRESTRIAL]];
+    const availableLayers = LAYERS_TOGGLE_CONFIG[biodiversityLayerVariant][TERRESTRIAL][resolution];
+    const layerTaxa = activeBiodiversityLayers.length ? activeBiodiversityLayers[0].slice(0, activeBiodiversityLayers[0].indexOf("-")) : ""
+    const hasMatchingLayer = availableLayers && availableLayers.find(layer => layer.value.includes(layerTaxa));
+    if (hasMatchingLayer) {
+      // select matching layer on selected variant
+      handleLayerToggle(hasMatchingLayer);
+    } else if(availableLayers) {
+      // select first element if there's no matching layer
+      handleLayerToggle(availableLayers[0]);
+    } else {
+      // select first element if there's no maching resolution
+      handleLayerToggle(deafaultResolutionLayers[0]);
+    }
+
+  }, [biodiversityLayerVariant])
+
   const handleTabSelection = (slug) => {
     changeUI({ biodiversityLayerVariant: slug });
   };
@@ -76,13 +101,6 @@ const BiodiversitySidebarCard = (props)  => {
       setSelectedLayer(option.layer);
     }
   }
-
-  // useSelectLayersOnTabChange({
-  //   biodiversityLayerVariant,
-  //   activeLayers,
-  //   biodiversityCategories,
-  //   handleClearAndAddLayers
-  // });
 
   return (
     <Component
