+++ src/constants/ui-params.js
@@ -14,7 +14,7 @@ export const LOCAL_SCENE_TABS = [
   {slug: LOCAL_SCENE_TABS_SLUGS.RANKING, title: 'ranking'},
 ]
 
-export const BIODIVERSITY_TABS_SLUGS = LAYER_VARIANTS;
+const BIODIVERSITY_TABS_SLUGS = LAYER_VARIANTS;
 
 export const BIODIVERSITY_DEFAULT_TAB = BIODIVERSITY_TABS_SLUGS.PRIORITY;
